requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0
pillow>=8.3.0
tqdm>=4.62.0

# Deep learning
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.10.0
timm>=0.5.0
albumentations>=1.0.0

# Image processing
opencv-python>=4.5.0
pdf2image>=1.16.0
pytesseract>=0.3.8

# Document processing
layoutparser>=0.3.0
donut-python>=1.0.0
easyocr>=1.5.0

# Text processing
nltk>=3.6.0
spacy>=3.1.0
regex>=2021.8.0

# Evaluation
editdistance>=0.5.0
levenshtein>=0.15.0
pycocoevalcap>=1.2.0

# Visualization
jupyter>=1.0.0
ipywidgets>=7.6.0
plotly>=5.3.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Documentation
sphinx>=4.1.0
sphinx-rtd-theme>=0.5.0
